# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EMARA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      responses:
        "200":
          description: List projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectWithClient" }
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /projects/{projectId}/files:
    get:
      operationId: listProjectFiles
      tags: [files]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FileAsset" }
    post:
      operationId: createProjectFile
      tags: [files]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFileBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileAsset" }
...
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /projects/{projectId}/payments:
    get:
      operationId: listProjectPayments
      tags: [payments]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }
    post:
      operationId: createPayment
      tags: [payments]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePaymentBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
...
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePaymentBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
    delete:
      operationId: deletePayment
      tags: [payments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    ProjectStatus:
      type: string
      enum:
        - new_lead
        - quotation_sent
        - contract_signed
        - site_survey
        - planning_2d
        - mood_board
        - design_3d
        - client_review
        - shop_drawings
        - boq
        - execution_ready
        - completed
        - cancelled
    ProjectType:
      type: string
      enum: [apartment, villa, office, restaurant, clinic, retail_store, other]
    DesignStyle:
      type: string
      enum: [modern, classic, neoclassical, minimal, japandi, industrial, contemporary, luxury]
    ScopeType:
      type: string
      enum: [design_only, design_shop_drawings, design_supervision, design_execution, turnkey]
    Project:
      type: object
      properties:
        id: { type: integer }
        clientId: { type: integer }
        projectName: { type: string }
        projectType: { $ref: "#/components/schemas/ProjectType" }
        location: { type: ["string", "null"] }
        areaSqm: { type: number }
        designStyle: { $ref: "#/components/schemas/DesignStyle" }
        scopeType: { $ref: "#/components/schemas/ScopeType" }
        budget: { type: ["number", "null"] }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        startDate: { type: ["string", "null"], format: date }
        expectedEndDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - clientId
        - projectName
        - projectType
        - location
        - areaSqm
        - designStyle
        - scopeType
        - budget
        - status
        - startDate
        - expectedEndDate
        - notes
        - createdAt
    ProjectWithClient:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            clientName: { type: string }
            progressPercent: { type: integer }
          required: [clientName, progressPercent]
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            clientName: { type: string }
            progressPercent: { type: integer }
            stagesTotal: { type: integer }
            stagesCompleted: { type: integer }
          required: [clientName, progressPercent, stagesTotal, stagesCompleted]
    CreateProjectBody:
      type: object
      properties:
        clientId: { type: integer }
        projectName: { type: string }
        projectType: { $ref: "#/components/schemas/ProjectType" }
        location: { type: ["string", "null"] }
        areaSqm: { type: number }
        designStyle: { $ref: "#/components/schemas/DesignStyle" }
        scopeType: { $ref: "#/components/schemas/ScopeType" }
        budget: { type: ["number", "null"] }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        startDate: { type: ["string", "null"], format: date }
        expectedEndDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
      required: [clientId, projectName, projectType, areaSqm, designStyle, scopeType, status]
    UpdateProjectBody:
      type: object
      properties:
        clientId: { type: integer }
        projectName: { type: string }
        projectType: { $ref: "#/components/schemas/ProjectType" }
        location: { type: ["string", "null"] }
        areaSqm: { type: number }
        designStyle: { $ref: "#/components/schemas/DesignStyle" }
        scopeType: { $ref: "#/components/schemas/ScopeType" }
        budget: { type: ["number", "null"] }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        startDate: { type: ["string", "null"], format: date }
        expectedEndDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
...
    FileCategory:
      type: string
      enum:
        - contract
        - site_survey
        - plan_2d
        - mood_board
        - render_3d
        - shop_drawing
        - boq
        - invoice
        - other
    FileAsset:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: ["string", "null"] }
        stageId: { type: ["integer", "null"] }
        uploadedById: { type: integer }
        uploadedByName: { type: ["string", "null"] }
        fileName: { type: string }
        fileUrl: { type: string }
        fileType: { type: ["string", "null"] }
        category: { $ref: "#/components/schemas/FileCategory" }
        version: { type: integer }
        isPublic: { type: boolean }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - projectId
        - projectName
        - stageId
        - uploadedById
        - uploadedByName
        - fileName
        - fileUrl
        - fileType
        - category
        - version
        - isPublic
        - notes
        - createdAt
    CreateFileBody:
      type: object
      properties:
        stageId: { type: ["integer", "null"] }
        fileName: { type: string }
        fileUrl: { type: string }
        fileType: { type: ["string", "null"] }
        category: { $ref: "#/components/schemas/FileCategory" }
        version: { type: integer }
        isPublic: { type: boolean }
        notes: { type: ["string", "null"] }
      required: [fileName, fileUrl, category, version, isPublic]
...
    PaymentStatus:
      type: string
      enum: [pending, paid, overdue, cancelled]
    PaymentType:
      type: string
      enum:
        - design_deposit
        - approval_2d
        - approval_mood
        - approval_3d
        - shop_drawing
        - boq
        - execution
        - final
        - other
    Payment:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        title: { type: string }
        amount: { type: number }
        dueDate: { type: ["string", "null"], format: date }
        paidDate: { type: ["string", "null"], format: date }
        status: { $ref: "#/components/schemas/PaymentStatus" }
        paymentType: { $ref: "#/components/schemas/PaymentType" }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - projectId
        - title
        - amount
        - dueDate
        - paidDate
        - status
        - paymentType
        - notes
        - createdAt
...
    CreatePaymentBody:
      type: object
      properties:
        title: { type: string }
        amount: { type: number }
        dueDate: { type: ["string", "null"], format: date }
        paymentType: { $ref: "#/components/schemas/PaymentType" }
        notes: { type: ["string", "null"] }
      required: [title, amount, paymentType]
    UpdatePaymentBody:
      type: object
      properties:
        title: { type: string }
        amount: { type: number }
        dueDate: { type: ["string", "null"], format: date }
        paidDate: { type: ["string", "null"], format: date }
        status: { $ref: "#/components/schemas/PaymentStatus" }
        paymentType: { $ref: "#/components/schemas/PaymentType" }
        notes: { type: ["string", "null"] }
...